+++ src/ff7_data.h
@@ -518,7 +518,9 @@ inline void ff7_find_externals(struct ff7_game_obj* game_object)
 	ff7_externals.opcode_biton = common_externals.execute_opcode_table[0x82];
 	ff7_externals.opcode_pc = common_externals.execute_opcode_table[0xA0];
 	ff7_externals.opcode_kawai = common_externals.execute_opcode_table[0x28];
+	ff7_externals.opcode_ifub = common_externals.execute_opcode_table[0x14];
 
+	ff7_externals.field_opcode_14_sub_6117CB = (int(*)())get_relative_call(common_externals.execute_opcode_table[0x14], 0x4);
 	ff7_externals.field_opcode_08_sub_61D0D4 = get_relative_call(common_externals.execute_opcode_table[0x08], 0x5A);
 	ff7_externals.field_opcode_08_09_set_rotation_61DB2C = (void(*)(short, byte, byte))get_relative_call(ff7_externals.field_opcode_08_sub_61D0D4, 0x196);
 	ff7_externals.field_opcode_AA_2A_sub_616476 = get_relative_call(common_externals.execute_opcode_table[0xAA], 0x26);